backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
httpx==0.27.0  # Only if you use it in llm_client.py
pandas
python-multipart
openpyxl==3.1.2
langgraph
langchain
langchain-core
langchain-groq